Move the CachedGitSourceRepository to the Git functionality directory
A recent decision taken was to do the grouping by what domain it serves rather than grouping by what thing it is
