+++ Corgibytes.Freshli.Cli/Functionality/Git/ICachedGitSourceRepository.cs
@@ -1,7 +1,7 @@
 using System.IO;
 using Corgibytes.Freshli.Cli.DataModel;
 
-namespace Corgibytes.Freshli.Cli.Repositories;
+namespace Corgibytes.Freshli.Cli.Functionality.Git;
 
 public interface ICachedGitSourceRepository
 {
